+++ mne_bids/copyfiles.py
@@ -20,6 +20,7 @@ from mne.io import anonymize_info, read_raw_bdf, read_raw_brainvision, read_raw_
 from mne.utils import logger, verbose
 from scipy.io import loadmat, savemat
 
+from mne_bids._fileio import _open_lock
 from mne_bids.path import BIDSPath, _mkdir_p, _parse_ext
 from mne_bids.utils import _check_anonymize, _get_mrk_meas_date, warn
 
@@ -63,7 +64,7 @@ def _get_brainvision_encoding(vhdr_file):
         in the header.
 
     """
-    with open(vhdr_file, "rb") as ef:
+    with _open_lock(vhdr_file, "rb") as ef:
         enc = ef.read()
         if enc.find(b"Codepage=") != -1:
             enc = enc[enc.find(b"Codepage=") + 9 :]
@@ -101,7 +102,7 @@ def _get_brainvision_paths(vhdr_path):
     enc = _get_brainvision_encoding(vhdr_path)
 
     # ..and read it
-    with open(vhdr_path, encoding=enc) as f:
+    with _open_lock(vhdr_path, encoding=enc) as f:
         lines = f.readlines()
 
     # Try to find data file .eeg/.dat
@@ -274,13 +275,14 @@ def copyfile_kit(src, dest, subject_id, session_id, task, run, _init_kwargs):
 def _replace_file(fname, pattern, replace):
     """Overwrite file, replacing end of lines matching pattern with replace."""
     new_content = []
-    for line in open(fname):
-        match = re.match(pattern, line)
-        if match:
-            line = match.group()[: -len(replace)] + replace + "\n"
-        new_content.append(line)
+    with _open_lock(fname, "r", encoding="utf-8") as fin:
+        for line in fin:
+            match = re.match(pattern, line)
+            if match:
+                line = match.group()[: -len(replace)] + replace + "\n"
+            new_content.append(line)
 
-    with open(fname, "w", encoding="utf-8") as fout:
+    with _open_lock(fname, "w", encoding="utf-8") as fout:
         fout.writelines(new_content)
 
 
@@ -395,15 +397,15 @@ def copyfile_brainvision(vhdr_src, vhdr_dest, anonymize=None, verbose=None):
         f"MarkerFile={basename_src}.vmrk",
     ]
 
-    with open(vhdr_src, encoding=enc) as fin:
-        with open(vhdr_dest, "w", encoding=enc) as fout:
+    with _open_lock(vhdr_src, encoding=enc) as fin:
+        with _open_lock(vhdr_dest, "w", encoding=enc) as fout:
             for line in fin.readlines():
                 if line.strip() in search_lines:
                     line = line.replace(basename_src, basename_dest)
                 fout.write(line)
 
-    with open(vmrk_file_path, encoding=enc) as fin:
-        with open(fname_dest + ".vmrk", "w", encoding=enc) as fout:
+    with _open_lock(vmrk_file_path, encoding=enc) as fin:
+        with _open_lock(fname_dest + ".vmrk", "w", encoding=enc) as fout:
             for line in fin.readlines():
                 if line.strip() in search_lines:
                     line = line.replace(basename_src, basename_dest)
@@ -513,7 +515,7 @@ def copyfile_edf(src, dest, anonymize=None, verbose=None):
             raise ValueError(f"Unsupported file type ({ext_src})")
 
         # Get subject info, recording info, and recording date
-        with open(dest, "rb") as f:
+        with _open_lock(dest, "rb") as f:
             f.seek(8)  # id_info field starts 8 bytes in
             id_info = f.read(80).decode("ascii").rstrip()
             rec_info = f.read(80).decode("ascii").rstrip()
@@ -540,7 +542,7 @@ def copyfile_edf(src, dest, anonymize=None, verbose=None):
         else:
             id_info = ["0", "X", "X", "X"]
             rec_info = ["Startdate", start_date, "X", "mne-bids_anonymize", "X"]
-        with open(dest, "r+b") as f:
+        with _open_lock(dest, "r+b") as f:
             f.seek(8)  # id_info field starts 8 bytes in
             f.write(bytes(" ".join(id_info).ljust(80), "ascii"))
             f.write(bytes(" ".join(rec_info).ljust(80), "ascii"))